Secure the home and shell for spark user (bsc#1087684)
The Ansible default behavior for creating a user is not very
secure.

As spark doesn't need a shell, specify /nonexistent home and
/usr/sbin/nologin shell for legacy install.
For suseified installs, just let the user creation from the .rpm
stand (which currently creates a home in /var/lib/spark and sets
the shell to /sbin/nologin).

Change-Id: I9e54692b7aad192b03832551f6292b0279c5d993