Only set log dir perms on legacy install (bsc#1094851)
The rpm creates the /var/log/spark directory with 750 perms and
spark:spark ownership.
One of the Jenkins post deployment checks noticed it was being
changed by the ansible scripts to 755. Investigation reveals this
is done so Beaver may access the log files to enable centralized
logging.

Separated the log dir creation into its own task and made it only
happen for legacy installs, and set 750 and spark:adm.

Also dropped the log/spark/events and /var/spark dirs as they are
not used.
Also moved the spark_conf_dir and spark_init_dir creation to the
install_legacy, as they are not needed in a rpm based deployment.

Change-Id: I02a1a3a699a60f22abe72eaed8c016de5443f7c4