Disable the REST API for the spark-master (bsc#1104760)
CVE-2018-11770 pointed out a vulnerability in Spark which could
allow a job to be submitted to the spark-master REST API without
authentication.  In SOC we block port 6066, so this is not a
critical problem.  However, to be thorough, we can disable the
REST API on the spark-master which is not used for monasca-transform.

Change-Id: Ifd09f5793d17ded5ff19f989f3b933ddc3e76c19